DVDInterface: slow down DVD seeking.
Increase seek delay from 1ms to 10ms; 10ms seems to be the right order of
magnitude according to some quick testing with Metroid Prime,

The DVDInterface code needs to be refactored to more realistically
represent how long it actually takes to respond to various commands;
adding 10ms to all commands (some of which don't involve accessing the
disc) is wrong.

+++ Source/Core/Core/CoreTiming.cpp
@@ -256,7 +256,7 @@ void AddEventToQueue(Event* ne)
 // This must be run ONLY from within the cpu thread
 // cyclesIntoFuture may be VERY inaccurate if called from anything else
 // than Advance
-void ScheduleEvent(int cyclesIntoFuture, int event_type, u64 userdata)
+void ScheduleEvent(unsigned long long cyclesIntoFuture, int event_type, u64 userdata)
 {
 	Event *ne = GetNewEvent();
 	ne->userdata = userdata;
