+++ Source/Core/Core/CoreTiming.h
@@ -41,7 +41,7 @@ void UnregisterAllEvents();
 
 // userdata MAY NOT CONTAIN POINTERS. userdata might get written and reloaded from disk,
 // when we implement state saves.
-void ScheduleEvent(int cyclesIntoFuture, int event_type, u64 userdata=0);
+void ScheduleEvent(unsigned long long cyclesIntoFuture, int event_type, u64 userdata=0);
 void ScheduleEvent_Threadsafe(int cyclesIntoFuture, int event_type, u64 userdata=0);
 void ScheduleEvent_Threadsafe_Immediate(int event_type, u64 userdata=0);
